Final contribution on v1
Integration test and webIntegrationTest.

+++ EmployeesContol/src/main/java/com/employees/control/Model/Employee.java
@@ -10,21 +10,21 @@ public class Employee {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private int id;
+	private Long id;
 	private String name;
 	
 	public Employee(){} 
 	
 	
-	public Employee(int id, String name) {
+	public Employee(Long id, String name) {
 		//super();
 		this.id = id;
 		this.name = name;
 	}
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
